algunos cambios

+++ NLayer.Negocios/ClienteServicios.cs
@@ -10,22 +10,28 @@ namespace NLayer.Negocios
 {
     public class ClienteServicios
     {
+        List<Cliente> listaclientes;
         public ClienteServicios()
         {
+            listaclientes = new List<Cliente>();
         }
-        public int IngresarCliente(Cliente cliente)
+        public int IngresarCliente(Cliente item)
         {
             //faltan validaciones de negocio.
-            TransactionResult resultado = ClienteMapper.Insert(cliente);
+            if (listaclientes.Any(o => o.Nombre == item.Nombre))
+            {
+                throw new ReservasException("El hotel ya existe");
+            }
+            TransactionResult resultado = ClienteMapper.Insert(item);
             if (resultado.IsOk)
                 return resultado.Id;
             else
                 return -1;
         }
-        public int ModificarCliente(Cliente cliente)
+        public int ModificarCliente(Cliente item)
         {
             //faltan validaciones de negocio.
-            TransactionResult resultado = ClienteMapper.Update(cliente);
+            TransactionResult resultado = ClienteMapper.Update(item);
             if (resultado.IsOk)
                 return resultado.Id;
             else
@@ -33,11 +39,16 @@ namespace NLayer.Negocios
         }
         public List<Cliente> TraerTodo()
         {
-            return ClienteMapper.TraerTodos();
+            listaclientes = ClienteMapper.TraerTodos();
+            return listaclientes;
         }
         public int EliminarCliente(int id)
         {
             //faltan validaciones de negocio.
+            if (!listaclientes.Any(o => o.Id == id))
+            {
+                throw new ReservasException("No se pudo encontrar el cliente");
+            }
             TransactionResult resultado = ClienteMapper.Delete(id);
             if (resultado.IsOk)
                 return resultado.Id;
