+++ NLayer.Negocios/HotelServicios.cs
@@ -10,12 +10,21 @@ namespace NLayer.Negocios
 {
     public class HotelServicios
     {
+        List<Hotel> listahoteles;
+        
         public HotelServicios()
         {
+            listahoteles = new List<Hotel>();
+            //listahoteles = HotelMapper.Hotel_getAll();
         }
         public int IngresarHotel(Hotel hotel)
         {
             //faltan validaciones de negocio.
+            if (listahoteles.Any(o => o.Nombre == hotel.Nombre))
+            {
+                throw new ReservasException("El hotel ya existe");
+            }
+
             TransactionResult resultado = HotelMapper.Insert(hotel);
             if (resultado.IsOk)
                 return resultado.Id;
@@ -33,11 +42,16 @@ namespace NLayer.Negocios
         }
         public List<Hotel> TraerTodo()
         {
-            return HotelMapper.Hotel_getAll();
+            listahoteles = HotelMapper.Hotel_getAll();
+            return listahoteles;
         }
         public int EliminarHotel(int id)
         {
             //faltan validaciones de negocio.
+            if (!listahoteles.Any(o => o.Id == id))
+            {
+                throw new ReservasException("No existe el hotel");
+            }
             TransactionResult resultado = HotelMapper.Delete(id);
             if (resultado.IsOk)
                 return resultado.Id;
